removing charts.hbs
I guess at some point I copied all of the content from chrats.hbs into
index.hbs because I want the chart project to be a drop-in inside the
data dissemination portal. Thus, the charts' index page is the charts
page and vice-versa. So we should be able to delete the charts.hbs
template with no ill effect. Right?